+++ test/plug/router_test.exs
@@ -20,6 +20,10 @@ defmodule Plug.RouterTest do
       conn |> resp(200, Enum.join(conn.script_name, ","))
     end
 
+    match "/params" do
+      conn |> resp(200, conn.params["param"])
+    end
+
     match "/throw", via: [:get, :post] do
       _ = conn
       throw :oops
@@ -58,6 +62,22 @@ defmodule Plug.RouterTest do
     forward "/step2", to: Forward
   end
 
+  defmodule SamplePlug do
+    import Plug.Conn
+
+    def init(options) do
+      options
+    end
+
+    def call(conn, []) do
+      send_resp(conn, 200, "ok")
+    end
+
+    def call(conn, options) do
+      send_resp(conn, 200, "#{inspect options}")
+    end
+  end
+
   defmodule Sample do
     use Plug.Router
     use Plug.ErrorHandler
@@ -72,6 +92,9 @@ defmodule Plug.RouterTest do
 
     get "/" do
       conn |> resp(200, "root")
+    end
+
+    get "/bodyless" do
     end
 
     get "/1/bar" do
@@ -106,15 +129,37 @@ defmodule Plug.RouterTest do
       conn |> resp(200, bat)
     end
 
+    plug = SamplePlug
+    opts = :foo
+    get "/plug/match", to: SamplePlug
+    get "/plug/match/options", to: plug, init_opts: opts
+
     forward "/step1", to: Reforward
     forward "/forward", to: Forward
     forward "/nested/forward", to: Forward
+
+    match "/params/get/:param" do
+      conn |> resp(200, conn.params["param"])
+    end
+
+    forward "/params/forward/:param", to: Forward
 
     get "/options/map", private: %{an_option: :a_value} do
       conn |> resp(200, inspect(conn.private))
     end
 
-    forward "/options/forward", to: Forward, private: %{an_option: :a_value}
+    get "/options/assigns", assigns: %{an_option: :a_value} do
+      conn |> resp(200, inspect(conn.assigns))
+    end
+
+    forward "/options/forward", to: Forward, private: %{an_option: :a_value},
+      assigns: %{another_option: :another_value}
+
+    plug = SamplePlug
+    opts = [foo: :bar]
+    forward "/plug/forward", to: SamplePlug
+    forward "/plug/options", to: SamplePlug, foo: :bar, private: %{baz: :qux}
+    forward "/plug/init_opts", to: plug, init_opts: opts, private: %{baz: :qux}
 
     match _ do
       conn |> resp(404, "oops")
@@ -134,6 +179,12 @@ defmodule Plug.RouterTest do
   test "dispatch root" do
     conn = call(Sample, conn(:get, "/"))
     assert conn.resp_body == "root"
+  end
+
+  test "dispatch empty body plug" do
+    assert_raise RuntimeError, ~r{expected dispatch/2 to return a Plug.Conn}, fn ->
+      call(Sample, conn(:get, "/bodyless"))
+    end
   end
 
   test "dispatch literal segment" do
@@ -197,6 +248,16 @@ defmodule Plug.RouterTest do
   test "dispatch wrong verb" do
     conn = call(Sample, conn(:post, "/1/bar"))
     assert conn.resp_body == "oops"
+  end
+
+  test "dispatch to plug" do
+    conn = call(Sample, conn(:get, "/plug/match"))
+    assert conn.resp_body == "ok"
+  end
+
+  test "dispatch to plug with options" do
+    conn = call(Sample, conn(:get, "/plug/match/options"))
+    assert conn.resp_body == ":foo"
   end
 
   test "dispatch with forwarding" do
@@ -317,16 +378,63 @@ defmodule Plug.RouterTest do
     end
   end
 
+  test "assigns path params to conn params and path_params" do
+    conn = call(Sample, conn(:get, "/params/get/a_value"))
+    assert conn.params["param"] == "a_value"
+    assert conn.path_params["param"] == "a_value"
+    assert conn.resp_body == "a_value"
+  end
+
+  test "assigns path params to conn params and path_params on forward" do
+    conn = call(Sample, conn(:get, "/params/forward/a_value/params"))
+    assert conn.params["param"] == "a_value"
+    assert conn.path_params["param"] == "a_value"
+    assert conn.resp_body == "a_value"
+  end
+
+  test "path params have priority over body and query params" do
+    conn = conn(:post, "/params/get/p_value", "param=b_value")
+    |> put_req_header("content-type", "application/x-www-form-urlencoded")
+    |> Plug.Parsers.call(Plug.Parsers.init(parsers: [:urlencoded]))
+
+    conn = call(Sample, conn)
+    assert conn.resp_body == "p_value"
+  end
+
   test "assigns route options to private conn map" do
     conn = call(Sample, conn(:get, "/options/map"))
     assert conn.private[:an_option] == :a_value
     assert conn.resp_body =~ ~s(an_option: :a_value)
   end
 
+  test "assigns route options to assigns conn map" do
+    conn = call(Sample, conn(:get, "/options/assigns"))
+    assert conn.assigns[:an_option] == :a_value
+    assert conn.resp_body =~ ~s(an_option: :a_value)
+  end
+
   test "assigns options on forward" do
     conn = call(Sample, conn(:get, "/options/forward"))
     assert conn.private[:an_option] == :a_value
+    assert conn.assigns[:another_option] == :another_value
     assert conn.resp_body == "forwarded"
+  end
+
+  test "forwards to a plug" do
+    conn = call(Sample, conn(:get, "/plug/forward"))
+    assert conn.resp_body == "ok"
+  end
+
+  test "forwards to a plug with options" do
+    conn = call(Sample, conn(:get, "/plug/options"))
+    assert conn.private[:baz] == :qux
+    assert conn.resp_body == "[foo: :bar]"
+  end
+
+  test "forwards to a plug with plug options" do
+    conn = call(Sample, conn(:get, "/plug/init_opts"))
+    assert conn.private[:baz] == :qux
+    assert conn.resp_body == "[foo: :bar]"
   end
 
   defp call(mod, conn) do